Restructure interfaces and types
