+++ src/btc-wallet.spec.ts
@@ -1,7 +1,7 @@
 import axios from "axios";
 import {BtcWallet} from "./btc-wallet";
 import * as assert from "assert";
-import {IAsyncEndpointResponse} from "./interfaces"
+import {IAsyncEndpointResponse} from "./interfaces/common";
 import {sandbox} from "./utils/spec-helpers";
 import {Waas} from "./waas";
 import {Wallet} from "./wallet";
