+++ src/eth-contract-wallet.spec.ts
@@ -1,7 +1,8 @@
 import * as assert from "assert";
 import axios from "axios"
 import {EthContractWallet} from "./eth-contract-wallet";
-import {IContractTransaction, IEthereumTransactionEstimation} from "./interfaces"
+import {IEthereumTransactionEstimation} from "./interfaces/ethereum";
+import {IContractTransaction} from "./interfaces/ethereum-contract";
 import {sandbox} from "./utils/spec-helpers";
 import {Waas} from "./waas";
 import {Wallet} from "./wallet";
