+++ src/eth.spec.ts
@@ -1,10 +1,11 @@
 import * as assert from "assert";
 import axios from "axios";
-import {SinonFakeTimers} from "sinon"
+import {SinonFakeTimers} from "sinon";
 import {MiningError, TimeoutError} from "./errors";
 import {isBitcoinMiningErrorData} from "./errors/mining-error";
 import {Ethereum} from "./eth";
-import {IBlockchainTransactionStatus, IEthereumTransactionStatus, ISearchTxResponse} from "./interfaces";
+import {IEthereumTransaction, ITransactionSearchResponse} from "./interfaces/ethereum";
+import {Transaction} from "./types/common";
 import {sandbox} from "./utils/spec-helpers";
 import {Waas} from "./waas";
 
@@ -59,7 +60,7 @@ describe("Ethereum", function() {
     // The method only calls the general function for wrapping search queries and this is tested in detail separately anyway.
     describe("getTransactions", function() {
         it("should execute the api call", async function() {
-            const sampleResponse: ISearchTxResponse = {
+            const sampleResponse: ITransactionSearchResponse = {
                 hits: {total: 4},
                 list: [
                     {
@@ -96,7 +97,7 @@ describe("Ethereum", function() {
             const stub = this.sandbox.stub()
             this.waas.instance.get = stub;
             const timer: SinonFakeTimers = this.sandbox.useFakeTimers({toFake: ["setInterval"]}); // fake the timer
-            const status: IBlockchainTransactionStatus = {isError: false, blockNr: 777, status: "confirmed"} as any
+            const status: Transaction = {isError: false, blockNr: 777, status: "confirmed"} as any
             stub.resolves(status);
 
             await Promise.all([
@@ -118,7 +119,7 @@ describe("Ethereum", function() {
                         if (isBitcoinMiningErrorData(r.txData)) {
                             throw new Error("invalid error type");
                         }
-                        const txData = (r.txData) as IEthereumTransactionStatus;
+                        const txData = (r.txData) as IEthereumTransaction;
                         assert.strictEqual(txData.isError, true);
                         assert.strictEqual(txData.blockNr, undefined);
                     }),
