+++ src/request.ts
@@ -1,5 +1,5 @@
 import * as t from "typeforce";
-import {IAsyncRequestStatus} from "./interfaces";
+import {IAsyncRequestStatus} from "./interfaces/common";
 import {poll} from "./utils/polling-helper";
 import {Waas} from "./waas";
 import {IWaasMethod} from "./waas-method";
