+++ src/waas.ts
@@ -5,9 +5,10 @@ import * as t from "typeforce";
 import {Bitcoin} from "./btc";
 import {AuthenticationError, ConflictError, GeneralError, MiningError, NotFoundError} from "./errors";
 import {Ethereum} from "./eth";
-import {BlockchainTransactionStatuses, IBlockchainTransactionStatus, IWaasError, WaasErrorResponse} from "./interfaces";
+import {IWaasError} from "./interfaces/common";
+import {Transaction, WaasErrorResponse} from "./types/common";
 import {limiter} from "./utils/limiter";
-import {Request} from "./request"
+import {Request} from "./request";
 import {Wallet} from "./wallet";
 import {poll} from "./utils/polling-helper";
 
@@ -87,11 +88,6 @@ export const ethereumRecipientType = t.compile({
     data: "?String",
 });
 
-export interface IWaitForTxStatus {
-    status: BlockchainTransactionStatuses;
-    response: IBlockchainTransactionStatus;
-}
-
 /**
  * Instantiates a new API interface. Multiple instances with different settings can run in parallel
  * @param options - api options
@@ -125,8 +121,8 @@ export class Waas {
      * @param [timeout] - if the statusGetterCall did not resolved during the timeout period (in ms) the function will reject
      * @param [ms] - milliseconds delay between api polling attempts
      */
-    public static async waitForTxStatus(statusGetterCall: () => Promise<IBlockchainTransactionStatus>, hash?: string, timeout = 20e3, ms = 400): Promise<IBlockchainTransactionStatus> {
-        const validate = (s: IBlockchainTransactionStatus) => {
+    public static async waitForTxStatus(statusGetterCall: () => Promise<Transaction>, hash?: string, timeout = 20e3, ms = 400): Promise<Transaction> {
+        const validate = (s: Transaction) => {
             switch (s.status) {
                 case "confirmed":
                     return true
@@ -138,7 +134,7 @@ export class Waas {
             }
         }
 
-        return poll<IBlockchainTransactionStatus>(statusGetterCall, validate, `transaction status ${hash}`, timeout, ms)
+        return poll<Transaction>(statusGetterCall, validate, `transaction status ${hash}`, timeout, ms)
     }
 
     public instance: AxiosInstance;
